+++ packages/terra-functional-testing/tests/jest/terra-cli/wdio/test-runner.test.js
@@ -52,7 +52,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         theme: 'terra-default-theme',
         updateScreenshots: true,
-        useSeleniumStandalonService: true,
+        useSeleniumStandaloneService: true,
       };
 
       await TestRunner.run(options);
@@ -177,7 +177,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         themes: ['terra-default-theme'],
         updateScreenshots: true,
-        useSeleniumStandalonService: true,
+        useSeleniumStandaloneService: true,
       });
 
       expect(TestRunner.run).toHaveBeenCalledWith({
@@ -196,7 +196,7 @@ describe('Test Runner', () => {
         suite: 'test-suite',
         theme: 'terra-default-theme',
         updateScreenshots: true,
-        useSeleniumStandalonService: true,
+        useSeleniumStandaloneService: true,
       });
     });
   });