+++ packages/terra-functional-testing/CHANGELOG.md
@@ -6,7 +6,7 @@
   * Update specPath in BaseCompare to replace `node_modules` with `tests/wdio`.
 
 * Added
-  * Added useSeleniumStandalonService option for using the standalone-chrome host instead of the selenium docker service when building in Jenkins.
+  * Added useSeleniumStandaloneService option for using the standalone-chrome host instead of the selenium docker service when building in Jenkins.
 
 ## 1.2.0 - (April 23, 2021)